+++ rosomaxa/src/hyper/dynamic_selective.rs
@@ -1,15 +1,25 @@
+#[cfg(test)]
+#[path = "../../tests/unit/hyper/dynamic_selective_test.rs"]
+mod dynamic_selective_test;
+
 use super::*;
-use crate::algorithms::math::relative_distance;
+use crate::algorithms::math::{relative_distance, Remedian};
 use crate::algorithms::mdp::*;
 use crate::utils::{compare_floats, Random};
 use crate::Timer;
 use hashbrown::HashMap;
 use std::cmp::Ordering;
+use std::hash::{Hash, Hasher};
 use std::sync::Arc;
+use std::time::Duration;
 
 /// A collection of heuristic operators.
 pub type HeuristicOperators<C, O, S> =
     Vec<(Arc<dyn HeuristicOperator<Context = C, Objective = O, Solution = S> + Send + Sync>, String)>;
+
+/// Specifies a median estimator used to track medians of heuristic running time.
+/// Its values can be used to penalize instant heuristic reward.
+type RemedianUsize = Remedian<usize, fn(&usize, &usize) -> Ordering>;
 
 /// An experimental dynamic selective hyper heuristic which selects inner heuristics
 /// based on how they work during the search. The selection process is modeled by
@@ -23,6 +33,7 @@ where
     heuristic_simulator: Simulator<SearchState>,
     initial_estimates: HashMap<SearchState, ActionEstimates<SearchState>>,
     action_registry: SearchActionRegistry<C, O, S>,
+    heuristic_median: RemedianUsize,
 }
 
 impl<C, O, S> HyperHeuristic for DynamicSelective<C, O, S>
@@ -38,6 +49,7 @@ where
     fn search(&mut self, heuristic_ctx: &Self::Context, solutions: Vec<&Self::Solution>) -> Vec<Self::Solution> {
         let registry = &self.action_registry;
         let estimates = &self.initial_estimates;
+        let median = &self.heuristic_median;
 
         let agents = solutions
             .into_iter()
@@ -47,24 +59,42 @@ where
                     original: solution,
                     registry,
                     estimates,
+                    median,
                     state: match compare_to_best(heuristic_ctx, solution) {
-                        Ordering::Greater => SearchState::Diverse,
-                        _ => SearchState::BestKnown,
+                        Ordering::Greater => SearchState::Diverse(Default::default()),
+                        _ => SearchState::BestKnown(Default::default()),
                     },
                     solution: Some(solution.deep_copy()),
+                    runtime: Vec::default(),
                 })
             })
             .collect();
 
-        let individuals = self
+        let (individuals, runtimes) = self
             .heuristic_simulator
             .run_episodes(agents, heuristic_ctx.environment().parallelism.clone(), |state, values| match state {
-                SearchState::BestKnown => values.iter().max_by(|a, b| compare_floats(**a, **b)).cloned().unwrap_or(0.),
+                SearchState::BestKnown { .. } => {
+                    values.iter().max_by(|a, b| compare_floats(**a, **b)).cloned().unwrap_or(0.)
+                }
                 _ => values.iter().sum::<f64>() / values.len() as f64,
             })
             .into_iter()
-            .filter_map(|agent| agent.solution)
-            .collect();
+            .filter_map(|agent| {
+                #[allow(clippy::manual_map)]
+                match agent.solution {
+                    Some(solution) => Some((solution, agent.runtime)),
+                    _ => None,
+                }
+            })
+            .fold((Vec::new(), Vec::new()), |mut acc, (solution, runtime)| {
+                acc.0.push(solution);
+                acc.1.extend(runtime.into_iter());
+                acc
+            });
+
+        runtimes.into_iter().for_each(|value| {
+            self.heuristic_median.add_observation(value.as_millis() as usize);
+        });
 
         try_exchange_estimates(&mut self.heuristic_simulator);
 
@@ -92,16 +122,47 @@ where
                 Box::new(EpsilonWeighted::new(0.1, random)),
             ),
             initial_estimates: vec![
-                (SearchState::BestKnown, operator_estimates.clone()),
-                (SearchState::Diverse, operator_estimates),
-                (SearchState::BestMajorImprovement, Default::default()),
-                (SearchState::BestMinorImprovement, Default::default()),
-                (SearchState::DiverseImprovement, Default::default()),
-                (SearchState::Stagnated, Default::default()),
+                (SearchState::BestKnown(Default::default()), operator_estimates.clone()),
+                (SearchState::Diverse(Default::default()), operator_estimates),
+                (SearchState::BestMajorImprovement(Default::default()), Default::default()),
+                (SearchState::BestMinorImprovement(Default::default()), Default::default()),
+                (SearchState::DiverseImprovement(Default::default()), Default::default()),
+                (SearchState::Stagnated(Default::default()), Default::default()),
             ]
             .into_iter()
             .collect(),
+            heuristic_median: RemedianUsize::new(11, |a, b| a.cmp(b)),
             action_registry: SearchActionRegistry { heuristics: operators },
+        }
+    }
+}
+
+#[derive(Default, Clone)]
+struct MedianRatio {
+    pub ratio: f64,
+}
+
+impl Hash for MedianRatio {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        0.hash(state)
+    }
+}
+
+impl PartialEq for MedianRatio {
+    fn eq(&self, _: &Self) -> bool {
+        true
+    }
+}
+
+impl Eq for MedianRatio {}
+
+impl MedianRatio {
+    pub fn eval(&self, value: f64) -> f64 {
+        match (self.ratio, compare_floats(value, 0.)) {
+            (ratio, _) if ratio < 1.001 => value,
+            (ratio, Ordering::Equal) => -2. * ratio,
+            (ratio, Ordering::Less) => value * ratio,
+            (ratio, Ordering::Greater) => value / ratio,
         }
     }
 }
@@ -109,17 +170,17 @@ where
 #[derive(PartialEq, Eq, Hash, Clone)]
 enum SearchState {
     /// A state with the best known solution.
-    BestKnown,
+    BestKnown(MedianRatio),
     /// A state with diverse (not the best known) solution.
-    Diverse,
+    Diverse(MedianRatio),
     /// A state with new best known solution found (major improvement).
-    BestMajorImprovement,
+    BestMajorImprovement(MedianRatio),
     /// A state with new best known solution found (minor improvement).
-    BestMinorImprovement,
+    BestMinorImprovement(MedianRatio),
     /// A state with improved diverse solution.
-    DiverseImprovement,
+    DiverseImprovement(MedianRatio),
     /// A state with equal or degraded solution.
-    Stagnated,
+    Stagnated(MedianRatio),
 }
 
 impl State for SearchState {
@@ -127,12 +188,12 @@ impl State for SearchState {
 
     fn reward(&self) -> f64 {
         match &self {
-            SearchState::BestKnown => 0.,
-            SearchState::Diverse => 0.,
-            SearchState::BestMajorImprovement => 1000.,
-            SearchState::BestMinorImprovement => 100.,
-            SearchState::DiverseImprovement => 10.,
-            SearchState::Stagnated => -1.,
+            SearchState::BestKnown(median_ratio) => median_ratio.eval(0.),
+            SearchState::Diverse(median_ratio) => median_ratio.eval(0.),
+            SearchState::BestMajorImprovement(median_ratio) => median_ratio.eval(1000.),
+            SearchState::BestMinorImprovement(median_ratio) => median_ratio.eval(100.),
+            SearchState::DiverseImprovement(median_ratio) => median_ratio.eval(10.),
+            SearchState::Stagnated(median_ratio) => median_ratio.eval(-1.),
         }
     }
 }
@@ -161,9 +222,11 @@ where
     heuristic_ctx: &'a C,
     registry: &'a SearchActionRegistry<C, O, S>,
     estimates: &'a HashMap<SearchState, ActionEstimates<SearchState>>,
+    median: &'a RemedianUsize,
     state: SearchState,
     original: &'a S,
     solution: Option<S>,
+    runtime: Vec<Duration>,
 }
 
 impl<'a, C, O, S> Agent<SearchState> for SearchAgent<'a, C, O, S>
@@ -181,13 +244,12 @@ where
     }
 
     fn take_action(&mut self, action: &<SearchState as State>::Action) {
-        let new_solution = match action {
+        let (new_solution, duration) = match action {
             SearchAction::Search { heuristic_idx } => {
                 let solution = self.solution.as_ref().unwrap();
                 let (heuristic, _) = &self.registry.heuristics[*heuristic_idx];
-                let (new_solution, _) = Timer::measure_duration(|| heuristic.search(self.heuristic_ctx, solution));
 
-                new_solution
+                Timer::measure_duration(|| heuristic.search(self.heuristic_ctx, solution))
             }
         };
 
@@ -195,6 +257,16 @@ where
 
         let compare_to_old = objective.total_order(&new_solution, self.original);
         let compare_to_best = compare_to_best(self.heuristic_ctx, &new_solution);
+
+        let ratio = MedianRatio {
+            ratio: self.median.approx_median().map_or(1., |median| {
+                if median == 0 {
+                    1.
+                } else {
+                    duration.as_millis() as f64 / median as f64
+                }
+            }),
+        };
 
         self.state = match (compare_to_old, compare_to_best) {
             (_, Ordering::Less) => {
@@ -210,16 +282,17 @@ where
                 );
 
                 if is_significant_change {
-                    SearchState::BestMajorImprovement
+                    SearchState::BestMajorImprovement(ratio)
                 } else {
-                    SearchState::BestMinorImprovement
+                    SearchState::BestMinorImprovement(ratio)
                 }
             }
-            (Ordering::Less, _) => SearchState::DiverseImprovement,
-            (_, _) => SearchState::Stagnated,
+            (Ordering::Less, _) => SearchState::DiverseImprovement(ratio),
+            (_, _) => SearchState::Stagnated(ratio),
         };
 
         self.solution = Some(new_solution);
+        self.runtime.push(duration)
     }
 }
 
@@ -227,8 +300,8 @@ fn try_exchange_estimates(heuristic_simulator: &mut Simulator<SearchState>) {
     let (best_known_max, diverse_state_max) = {
         let state_estimates = heuristic_simulator.get_state_estimates();
         (
-            state_estimates.get(&SearchState::BestKnown).and_then(|state| state.max_estimate()),
-            state_estimates.get(&SearchState::Diverse).and_then(|state| state.max_estimate()),
+            state_estimates.get(&SearchState::BestKnown(Default::default())).and_then(|state| state.max_estimate()),
+            state_estimates.get(&SearchState::Diverse(Default::default())).and_then(|state| state.max_estimate()),
         )
     };
 
@@ -238,8 +311,9 @@ fn try_exchange_estimates(heuristic_simulator: &mut Simulator<SearchState>) {
         diverse_state_max.map_or(false, |(_, max)| compare_floats(max, 0.) == Ordering::Greater);
 
     if is_best_known_stagnation && is_diverse_improvement {
-        let estimates = heuristic_simulator.get_state_estimates().get(&SearchState::Diverse).unwrap().clone();
-        heuristic_simulator.set_action_estimates(SearchState::BestKnown, estimates);
+        let estimates =
+            heuristic_simulator.get_state_estimates().get(&SearchState::Diverse(Default::default())).unwrap().clone();
+        heuristic_simulator.set_action_estimates(SearchState::BestKnown(Default::default()), estimates);
     }
 }